# solutions/issuedesk-web/static/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "issuedesk API — API-key endpoints",
    "version": "0.1.0",
    "description": "The subset of the issuedesk API that an **API key** can call.\n\nAPI keys are read-only — they may call any `GET` endpoint — **with one family of exceptions**: managing comments. A key may post (`POST …/comments`), edit (`PATCH …/comments/{id}`) and delete (`DELETE …/comments/{id}`) comments. Everything else (sign-in, creating/editing/deleting issues, user and key management) is reserved for the browser app and is not shown here.\n\n## Authenticating\n\nSend your key as `X-API-Key: idk_…` (or `Authorization: Bearer idk_…`). API keys are exempt from the browser-lock, so curl/agents need no extra headers. A key acts on behalf of the user who created it, so any comment it posts is attributed to that user; it may edit or delete its own comments (and, as it carries admin rights, any other comment).\n\nThe same endpoints also work from the signed-in browser app via a JWT bearer token."
  },
  "servers": [{ "url": "/", "description": "This host" }],
...
  "components": {
    "securitySchemes": {
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "An issuedesk API key (idk_…)."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "A JWT from the signed-in browser app (also accepted on these endpoints)."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "example": "not_found" },
          "message": { "type": "string", "example": "issue not found" }
        }
      },
      "Label": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "color": { "type": "string", "example": "#22c55e" }
        }
      },
      "IssueDetail": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "projectId": { "type": "string", "format": "uuid" },
          "projectKey": { "type": "string", "example": "WAT" },
          "key": { "type": "string", "example": "WAT-1" },
          "number": { "type": "integer", "format": "int64" },
          "title": { "type": "string" },
          "description": { "type": "string", "nullable": true, "description": "Rich HTML/Markdown." },
          "type": { "type": "integer", "description": "0 bug, 1 task, 2 story, 3 epic" },
          "status": { "type": "integer", "description": "0 Todo, 1 InProgress, 2 Implemented, 3 InReview, 4 Done" },
          "priority": { "type": "integer", "description": "0 low, 1 medium, 2 high, 3 urgent" },
          "assigneeId": { "type": "string", "format": "uuid", "nullable": true },
          "assigneeName": { "type": "string", "nullable": true },
          "reporterId": { "type": "string", "format": "uuid" },
          "reporterName": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "labels": { "type": "array", "items": { "$ref": "#/components/schemas/Label" } }
        }
      },
...
      "Comment": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "issueId": { "type": "string", "format": "uuid" },
          "authorId": { "type": "string", "format": "uuid" },
          "authorName": { "type": "string" },
          "body": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "CreateCommentRequest": {
        "type": "object",
        "required": ["body"],
        "properties": { "body": { "type": "string", "minLength": 1 } }
      },
      "UpdateCommentRequest": {
        "type": "object",
        "required": ["body"],
        "properties": { "body": { "type": "string", "minLength": 1 } }
      },
      "OkResponse": {
        "type": "object",
        "properties": { "ok": { "type": "boolean", "example": true } }
      },
      "Activity": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "actorName": { "type": "string" },
          "action": { "type": "integer" },
          "oldValue": { "type": "string", "nullable": true },
          "newValue": { "type": "string", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "IssueLink": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "linkType": { "type": "integer", "description": "0 relates to, 1 blocks, 2 is blocked by, 3 duplicates, 4 is duplicated by" },
          "issueId": { "type": "string", "format": "uuid" },
          "key": { "type": "string" },
          "number": { "type": "integer", "format": "int64" },
          "title": { "type": "string" },
          "status": { "type": "integer" },
          "projectKey": { "type": "string" }
        }
      },
      "TicketBundle": {
        "type": "object",
        "properties": {
          "issue": { "$ref": "#/components/schemas/IssueDetail" },
          "comments": { "type": "array", "items": { "$ref": "#/components/schemas/Comment" } },
          "activity": { "type": "array", "items": { "$ref": "#/components/schemas/Activity" } },
          "links": { "type": "array", "items": { "$ref": "#/components/schemas/IssueLink" } },
          "attachments": { "type": "array", "items": { "$ref": "#/components/schemas/Attachment" } }
        }
      },
      "Attachment": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "issueId": { "type": "string", "format": "uuid" },
          "filename": { "type": "string" },
          "storedPath": { "type": "string" },
          "sizeBytes": { "type": "integer", "format": "int64" },
          "mimeType": { "type": "string" },
          "uploadedBy": { "type": "string", "format": "uuid" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  },
  "security": [{ "apiKeyAuth": [] }, { "bearerAuth": [] }],
  "paths": {
    "/api/tickets/{slug}": {
      "get": {
        "tags": ["tickets"],
        "summary": "Fetch a whole ticket by public key (e.g. WAT-1)",
        "description": "The URL-friendly relay entry point: an API key plus the key from the ticket URL is enough. `?format=md` returns Markdown; otherwise JSON.",
        "parameters": [
          { "name": "slug", "in": "path", "required": true, "schema": { "type": "string" }, "example": "WAT-1" },
          { "name": "format", "in": "query", "required": false, "schema": { "type": "string", "enum": ["md"] }, "description": "`md` for Markdown; omit for JSON." }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/TicketBundle" } },
              "text/markdown": { "schema": { "type": "string" } }
            }
          },
          "404": { "description": "Not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/api/tickets/{slug}/comments": {
      "post": {
        "tags": ["tickets"],
        "summary": "Add a comment to a ticket by public key",
        "description": "Slug-addressed counterpart to POST /api/issues/{id}/comments, so a relay client (API key) can comment back with only the key from the ticket URL. The comment is attributed to the key's creator.",
        "parameters": [
          { "name": "slug", "in": "path", "required": true, "schema": { "type": "string" }, "example": "WAT-1" }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateCommentRequest" } } }
        },
        "responses": {
          "200": { "description": "Created comment", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Comment" } } } },
          "404": { "description": "Not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/api/tickets/{slug}/comments/{commentId}": {
      "patch": {
        "tags": ["tickets"],
        "summary": "Edit a comment on a ticket by public key",
        "description": "Slug-addressed counterpart to PATCH /api/comments/{id}. The caller must own the comment (or be an admin; API keys carry admin rights).",
        "parameters": [
          { "name": "slug", "in": "path", "required": true, "schema": { "type": "string" }, "example": "WAT-1" },
          { "name": "commentId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpdateCommentRequest" } } }
        },
        "responses": {
          "200": { "description": "Updated comment", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Comment" } } } },
          "403": { "description": "Not your comment", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "404": { "description": "Not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      },
      "delete": {
        "tags": ["tickets"],
        "summary": "Delete a comment on a ticket by public key",
        "description": "Slug-addressed counterpart to DELETE /api/comments/{id}. The caller must own the comment (or be an admin; API keys carry admin rights).",
        "parameters": [
          { "name": "slug", "in": "path", "required": true, "schema": { "type": "string" }, "example": "WAT-1" },
          { "name": "commentId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": { "description": "Deleted", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OkResponse" } } } },
          "403": { "description": "Not your comment", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "404": { "description": "Not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/api/tickets/{slug}/attachments": {
      "get": {
        "tags": ["tickets"],
        "summary": "List a ticket's attachments by public key",
        "description": "Slug-addressed counterpart to GET /api/issues/{id}/attachments. Each item's `id` downloads via GET /api/attachments/{id}.",
        "parameters": [
          { "name": "slug", "in": "path", "required": true, "schema": { "type": "string" }, "example": "WAT-1" }
        ],
        "responses": {
          "200": { "description": "OK", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Attachment" } } } } },
          "404": { "description": "Not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      },
      "post": {
        "tags": ["tickets"],
        "summary": "Upload an attachment to a ticket by public key",
        "description": "Slug-addressed counterpart to POST /api/issues/{id}/attachments, so a relay client (API key) can attach a file when commenting back with only the key from the ticket URL. Send multipart/form-data with a single file field. The returned `id` downloads via GET /api/attachments/{id}.",
        "parameters": [
          { "name": "slug", "in": "path", "required": true, "schema": { "type": "string" }, "example": "WAT-1" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["file"],
                "properties": { "file": { "type": "string", "format": "binary" } }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Created attachment", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Attachment" } } } },
          "404": { "description": "Not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "413": { "description": "File exceeds the upload size limit", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
    "/api/issues/{id}/comments": {
      "get": {
        "tags": ["comments"],
        "summary": "List comments on an issue",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Comment" } } } } } }
      },
      "post": {
        "tags": ["comments"],
        "summary": "Add a comment to an issue (by id)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateCommentRequest" } } } },
        "responses": { "200": { "description": "Created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Comment" } } } } }
      }
    },
    "/api/comments/{id}": {
      "patch": {
        "tags": ["comments"],
        "summary": "Edit a comment (by id)",
        "description": "The caller must own the comment (or be an admin; API keys carry admin rights).",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpdateCommentRequest" } } } },
        "responses": {
          "200": { "description": "Updated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Comment" } } } },
          "403": { "description": "Not your comment", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "404": { "description": "Not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      },
      "delete": {
        "tags": ["comments"],
        "summary": "Delete a comment (by id)",
        "description": "The caller must own the comment (or be an admin; API keys carry admin rights).",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": {
          "200": { "description": "Deleted", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/OkResponse" } } } },
          "403": { "description": "Not your comment", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "404": { "description": "Not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
  }
}